Fix JSON writing and parsing with other locales
Certain locales use different characters to separate decimals in
floating point numbers. This can lead to invalid JSON files being
generated, when the decimals are separated by a comma. To fix this,
an invariant locale should be used for writing and parsing floating
point numbers.

This issue was originally noticed by @Deatrathias (Virtual_Deat on
Twitter), who also came up with a way to solve it. This commit is based
on his solution.